apps/api/pyproject.toml
[project]
name = "aegis-api"
version = "0.1.0"
description = "AEGIS FastAPI backend for disaster assessment"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.1",
    "supabase>=2.7.0",
    "ollama>=0.3.0",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"
